chore: Remove all references to WithBuiltIn
